Improve run.sh to automatically restart frida when needed (Checks every 5 seconds)

+++ run.sh
@@ -2,5 +2,16 @@
 # Source: https://frida.re/docs/android/
 
 adb root
-adb shell 'kill -9 $(pidof frida-server)'  # Clean up any previous processes
-adb shell "/data/local/tmp/frida-server &" &
+
+if adb shell "pidof frida-server" ; then
+  # Clean up any previous processes
+  adb shell 'kill -9 $(pidof frida-server)'
+fi
+
+while true; do
+  if ! adb shell "pidof frida-server" > /dev/null ; then
+    echo "Restarting frida-server"
+    adb shell "/data/local/tmp/frida-server &" &
+  fi
+  sleep 5
+done
